Larastan clean 3 (#37)
* Update UserActivityService.php - Parameter #2 $ttl of static method Illuminate\Support\Facades\Cache::remember() expects

* Update FeedService.php - Match expression does not handle remaining value: mixed

* Update AdminSetting.php - Expression on left side of ?? is not nullable.

* Update SettingsController.php - Ternary operator condition is always true.

+++ app/Services/UserActivityService.php
@@ -15,7 +15,7 @@ class UserActivityService
         $ymd = $ts->format('Ymd');
         $key = "ua:{$user->id}:{$ymd}";
 
-        $ttl = $ts->startOfDay()->addDay()->diffInSeconds($ts) + 5;
+        $ttl = (int) ($ts->startOfDay()->addDay()->diffInSeconds($ts) + 5);
 
         Cache::remember($key, $ttl, function () use ($ts, $user) {
             $user->forceFill(['last_active_at' => $ts])->saveQuietly();
